YA: moving public misc functions to auxiliar
